status: polish status indicators
Apparently there are two indicators; the problem is with users coming
online -- the primary status indicator stays the default color, and the
secondary one is transparent, resulting in the default indicator colors.

It's possible to brute-force it, and have custom colors of the secondary
status indicator apply by switching between servers, but that's more of
a hack and not necessarily a solution.

The solution is to overwite primary status indicator, so that when the
secondary one becomes transparent, it still uses custom colors. This way
things work as intended.